create main product
