+++ web/carousel/src/components/Account/VendorAddProduct/MainProductForm/MainProductForm.js
@@ -1,7 +1,11 @@
 import React, { Component } from "react";
 import classes from "../AddProduct.module.css";
-import { Button, Form, Input, Upload, Modal } from "antd";
+import { Button, Form, Input, Space } from "antd";
 import PicturesWall from "../../../PicturesWall";
+import { MinusCircleOutlined, PlusOutlined } from "@ant-design/icons";
+import ButtonSecondary from "../../../UI/ButtonSecondary/ButtonSecondary";
+
+const TOKEN = localStorage.getItem("token");
 
 const layout = {
   labelCol: { span: 8 },
@@ -21,55 +25,40 @@ const validateMessages = {
 
 class MainProductForm extends Component {
   render() {
-    // let parameterInputs = "";
-    // for (let i = 0; i < this.props.parameterInputs.length; i++) {}
-    // parameterInputs = this.props.parameterInputs.map((param, igKey) => {
-    //   console.log(param);
-    //   return (
-    //     <Form.Item
-    //       name={["user", "parameter_" + param]}
-    //       label={param.charAt(0).toUpperCase() + param.slice(1)}
-    //     >
-    //       <Input />
-    //     </Form.Item>
-    //   );
-    // });
-
     return (
       <Form
         {...layout}
         className={classes.myReset}
-        name="nest-messages"
-        // onFinish={this.props.clicked}
+        name="dynamic_form_nest_item"
+        onFinish={this.props.clicked}
         validateMessages={validateMessages}
       >
         <Form.Item
           name={["user", "title"]}
           label="Main Product Title"
-          // initialValue={this.props.product.title}
+          rules={[{ required: true }]}
         >
-          <Input disabled />
+          <Input placeholder="example: Iphone 8" />
         </Form.Item>
         <Form.Item
           name={["user", "description"]}
           label="Main Product Description"
-          // initialValue={this.props.product.title}
         >
           <Input />
         </Form.Item>
         <Form.Item
           name={["user", "brand"]}
           label="Brand"
-          // initialValue={this.props.product.brand}
+          rules={[{ required: true }]}
         >
-          <Input disabled />
+          <Input placeholder="example: Apple" />
         </Form.Item>
         <Form.Item
           name={["user", "category"]}
           label="Category"
-          // initialValue={this.props.product.category}
+          rules={[{ required: true }]}
         >
-          <Input disabled />
+          <Input placeholder="example: electronics" />
         </Form.Item>
 
         <Form.Item
@@ -79,16 +68,62 @@ class MainProductForm extends Component {
         >
           <Input placeholder="tag1, tag2, ..." />
         </Form.Item>
-        <Form.Item name={["user", "parameters"]} label="Parameters">
-          <div style={{ height: "20px" }} />
-          parameterInputs
+        <Form.Item label="Parameters">
+          <Form.List name={["user", "parameters"]}>
+            {(fields, { add, remove }) => (
+              <>
+                {fields.map((field) => (
+                  <Space
+                    key={field.key}
+                    style={{ display: "flex", marginBottom: 8 }}
+                    align="baseline"
+                  >
+                    <Form.Item
+                      {...field}
+                      name={[field.name, "name"]}
+                      fieldKey={[field.fieldKey, "name"]}
+                      rules={[
+                        { required: true, message: "Missing parameter name" },
+                      ]}
+                    >
+                      <Input placeholder="color" />
+                    </Form.Item>
+                    <Form.Item
+                      {...field}
+                      name={[field.name, "values"]}
+                      fieldKey={[field.fieldKey, "values"]}
+                      rules={[
+                        { required: true, message: "Missing parameter values" },
+                      ]}
+                    >
+                      <Input placeholder="white, black, ..." />
+                    </Form.Item>
+                    <MinusCircleOutlined onClick={() => remove(field.name)} />
+                  </Space>
+                ))}
+                <Form.Item>
+                  <Button
+                    type="dashed"
+                    onClick={() => add()}
+                    block
+                    icon={<PlusOutlined />}
+                  >
+                    Add Parameters
+                  </Button>
+                </Form.Item>
+              </>
+            )}
+          </Form.List>
         </Form.Item>
-        <Form.Item>
+        <Form.Item name={["user", "images"]}>
           <PicturesWall />
         </Form.Item>
         <Form.Item wrapperCol={{ ...layout.wrapperCol, offset: 8 }}>
           <Button type="primary" htmlType="submit">
             Submit
+          </Button>
+          <Button htmlType="button" onClick={this.props.onClick}>
+            Cancel
           </Button>
         </Form.Item>
       </Form>
