+++ web/carousel/src/components/Account/VendorAddProduct/MainProductTable/MainProductTable.js
@@ -7,34 +7,6 @@ import services from "../../../../apis/services";
 import MainProductForm from "../MainProductForm/MainProductForm";
 
 class MainProductTable extends Component {
-  postMainProduct = () => {
-    let url = "/mainProduct/";
-    let payload = {
-      title: "Nice TV ",
-      parameters: [
-        { name: "size", values: ["XL", "L", "M"] },
-        { name: "color", values: ["blue", "green", "yellow"] },
-      ],
-      description: "Description",
-      rating: 0,
-      numberOfRating: 0,
-      brand: "Samsung",
-      soldAmount: 0,
-      category: "electronics",
-      isConfirmed: false,
-      tags: ["Apple", "blue", "green", "yellow", "XL", "L", "M"],
-    };
-
-    services
-      .post(url, payload)
-      .then((response) => {
-        console.log(response);
-      })
-      .catch((error) => {
-        console.log("ERROR: " + error);
-      });
-  };
-
   render() {
     const columns = [
       {
